requirements.txt
webrtcvad==2.0.10
openai==1.3.0
pyaudio==0.2.11
openai-whisper==20231117
PyPDF2==3.0.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
sentence-transformers==2.2.2